+++ source/tests/Test-BlockChannelEmails.ps1
@@ -5,16 +5,14 @@ function Test-BlockChannelEmails {
         # Aligned
         # Parameters can be added here if needed
     )
-
     begin {
         # Dot source the class script if necessary
         #. .\source\Classes\CISAuditResult.ps1
         # Initialization code, if needed
         $recnum = "8.1.2"
+        Write-Verbose "Running Test-BlockChannelEmails for $recnum..."
     }
-
     process {
-
         try {
             # 8.1.2 (L1) Ensure users can't send emails to a channel email address
             #
@@ -31,11 +29,9 @@ function Test-BlockChannelEmails {
             #   - Condition A: The `AllowEmailIntoChannel` setting in Teams is not set to `False`.
             #   - Condition B: The setting `Users can send emails to a channel email address` is not set to `Off` in the Teams admin center.
             #   - Condition C: Verification using PowerShell indicates that the `AllowEmailIntoChannel` setting is enabled.
-
             # Retrieve Teams client configuration
             $teamsClientConfig = Get-CISMSTeamsOutput -Rec $recnum
             $allowEmailIntoChannel = $teamsClientConfig.AllowEmailIntoChannel
-
             # Prepare failure reasons and details based on compliance
             $failureReasons = if ($allowEmailIntoChannel) {
                 "Emails can be sent to a channel email address" # Condition A Fail: AllowEmailIntoChannel is True
@@ -43,14 +39,12 @@ function Test-BlockChannelEmails {
             else {
                 "N/A" # Condition A Pass: AllowEmailIntoChannel is False
             }
-
             $details = if ($allowEmailIntoChannel) {
                 "AllowEmailIntoChannel is set to True" # Condition B Fail: Emails are allowed
             }
             else {
                 "AllowEmailIntoChannel is set to False" # Condition B Pass: Emails are blocked
             }
-
             # Create and populate the CISAuditResult object
             $params = @{
                 Rec            = $recnum
@@ -66,7 +60,6 @@ function Test-BlockChannelEmails {
             $auditResult = Get-TestError -LastError $LastError -recnum $recnum
         }
     }
-
     end {
         # Return the audit result
         return $auditResult