+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on and uses the types of changes according to [Keep a Change
 
 ### Added
 
+- Added `ApprovedCloudStorageProviders` parameter to `Invoke-M365SecurityAudit` to allow for testing of approved cloud storage providers for 8.1.1.
+- Added `ApprovedFederatedDomains` parameter to `Invoke-M365SecurityAudit` to allow for testing of approved federated domains for 8.5.1.
+
+### Fixed
+
+- Fixed various MSTeams tests to be more accurate and include more properties in the output.
+
+## [0.1.18] - 2024-06-29
+
+### Added
+
 - Added `Get-PhishPolicyDetail` and `Test-PhishPolicyCompliance` private functions to help test for phishing policy compliance.
 
 ### Fixed
@@ -15,8 +26,6 @@ The format is based on and uses the types of changes according to [Keep a Change
 #### Changed
 
 - Changed main function parameter for Domain to `DomainName`.
-
-
 
 ## [0.1.17] - 2024-06-28